backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Data science and analysis
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.1
lightgbm==4.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# AI and LLM - OpenRouter only (NO OpenAI)
python-dotenv==1.0.0

# File processing
openpyxl==3.1.2
xlrd==2.0.1
PyPDF2==3.0.1
python-docx==1.1.0

# OCR capabilities
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Statistical analysis
statsmodels==0.14.0

# Utilities
requests==2.31.0
aiofiles==23.2.1

# Database (for conversation memory)
sqlalchemy==2.0.23

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 